ListOfPost page - Component separation
