+++ src/components/Header.jsx
@@ -1,5 +1,3 @@
-// Header.jsx
-
 import React from 'react';
 import { Link, useLocation } from 'react-router-dom';
 import titleImage from '../assets/images/title.png';
