+++ src/pages/SignUpPage.jsx
@@ -106,9 +106,6 @@ class SignUpPage extends React.Component {
                             disabled={!this.state.isValid}
                             text="회원가입"
                         />
-                        {/* <a href="sign-in">
-                            <p><small>로그인하러 가기</small></p>
-                        </a> */}
                         <HyperlinkText to="/signin" text="로그인하러 가기" />
                     </form>
                 </section>
